requirements.txt
pandas
numpy
matplotlib
torch==1.12.0
pyyaml
lxml
tqdm
yfinance